+++ src/app/[locale]/marketplace/_components/catalogConfiguration.ts
